requirements.txt
questionary>=2.0.0
rich>=13.0.0
pyyaml>=6.0
